log_aggregator/requirements.txt
django==5.1
python-dotenv==1.0.1
psycopg[binary,pool]==3.1.19
djangorestframework==3.15.2
django-filter==24.3
setuptools==73.0.1
drf-yasg==1.21.7
uWSGI==2.0.26
